# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CIS Phase2 - Crowdsourced Ideation API
  description: |
    Complete API for collaborative brainstorming system with voting capabilities.
    Enables creating topics, publishing ideas, and voting on them.
    
    **Version 1 (V1):** Uses MySQL for persistence. Stable.
    **Version 2 (V2):** Uses MongoDB for persistence. Experimental (US 1.1).
    
    **Business Rules:**
    - A user can vote only once per idea (unique vote)
    - Votes can only be cast when the topic is OPEN
    - Users can change/remove their vote while the topic is OPEN
    - When a topic is CLOSED, the winning idea is determined automatically based on vote count.
    - Once a topic is CLOSED, it cannot be reopened.
    - Ownership rules apply to updating and deleting topics, ideas, and votes.
    - Deleting a topic cascades and deletes all related ideas and votes.
  version: 2.1.0
  contact:
    name: CIS Phase2 Team
  license:
    name: MIT
servers:
  - url: http://localhost:5257
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Phase 1 User Management API
  parameters:
    TopicId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Topic ID (UUID string)
      example: "550e8400-e29b-41d4-a716-446655440000"
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Page number (starting from 0)
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        default: 10
      description: Number of items per page
    SortOrderParam:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: Sort order
  responses:
    BadRequest:
      description: Invalid data or business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_paging:
              summary: Invalid paging/sorting/filtering parameter
              value:
                error: "size must be >= 1."
    Unauthorized:
      description: Missing, invalid, or expired JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing_or_invalid_token:
              summary: Unauthorized (missing/invalid/expired token)
              value:
                error: "Unauthorized - Valid token required"
    Forbidden:
      description: Access denied due to ownership or business rules (e.g. closed topic)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            ownership:
              summary: User is not the owner of the resource
              value:
                error: "You are not authorized to modify this topic."
            closed_topic:
              summary: Business rule violation (topic is closed)
              value:
                error: "This topic is closed. No modifications allowed."
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            not_found:
              summary: Resource does not exist
              value:
                error: "Resource not found."
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Detailed error message"
        message:
          type: string
          example: "Error summary"
        errorCode:
          type: string
          nullable: true
          example: "DUPLICATE_VOTE"
      additionalProperties: true

    # ----- HATEOAS Link -----
    Link:
      type: object
      properties:
        href:
          type: string
          example: "api/v1/topics/550e8400-e29b-41d4-a716-446655440000"
        method:
          type: string
          example: "GET"
        rel:
          type: string
          example: "self"

    # ----- Topic Models -----
    TopicStatus:
      type: string
      enum:
        - OPEN
        - CLOSED
      description: |
        - OPEN: Ideas can be created and votes can be cast
        - CLOSED: No more ideas/votes accepted, winning idea is displayed
      example: OPEN
    WinningIdeaResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        topicId:
          type: string
        ownerId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isWinning:
          type: boolean
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    TopicResponse:
      type: object
      required:
        - id
        - title
        - status
        - ownerId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          maxLength: 200
          example: "Dashboard improvements"
        description:
          type: string
          nullable: true
          example: "Proposals to optimize the interface"
        status:
          $ref: '#/components/schemas/TopicStatus'
        ownerId:
          type: string
          description: Creator user ID
          example: "550e8400-e29b-41d4-a716-446655440001"
        createdAt:
          type: string
          format: date-time
          example: "2026-03-30T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-03-30T15:30:00Z"
        winningIdea:
          $ref: '#/components/schemas/WinningIdeaResponse'
          nullable: true
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
          example:
            - href: "api/v1/topics/550e8400-e29b-41d4-a716-446655440000"
              method: "GET"
              rel: "self"
            - href: "api/v1/ideas/topic/550e8400-e29b-41d4-a716-446655440000"
              method: "GET"
              rel: "ideas"
            - href: "api/v1/topics/550e8400-e29b-41d4-a716-446655440000"
              method: "PUT"
              rel: "update"
            - href: "api/v1/topics/550e8400-e29b-41d4-a716-446655440000"
              method: "DELETE"
              rel: "delete"
    PagedTopicResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TopicResponse'
        currentPage:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
    CreateTopicRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "New mobile features"
        description:
          type: string
          nullable: true
          example: "Share ideas about the mobile app"
      additionalProperties: false
    UpdateTopicRequest:
      type: object
      required:
        - title
        - status
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "New mobile features - Updated"
        description:
          type: string
          nullable: true
          example: "Focus on UX/UI"
        status:
          $ref: '#/components/schemas/TopicStatus'
      additionalProperties: false

    # ----- Idea Models -----
    IdeaResponse:
      type: object
      required:
        - id
        - topicId
        - ownerId
        - title
        - description
        - createdAt
        - updatedAt
        - isWinning
      properties:
        id:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440000"
        topicId:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        ownerId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        title:
          type: string
          example: "Implement dark mode"
        description:
          type: string
          example: "Offer a dark theme for all users"
        createdAt:
          type: string
          format: date-time
          example: "2026-03-30T11:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-03-30T11:30:00Z"
        isWinning:
          type: boolean
          example: false
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
          example:
            - href: "api/v1/ideas/660e8400-e29b-41d4-a716-446655440000"
              method: "GET"
              rel: "self"
            - href: "api/v1/topics/550e8400-e29b-41d4-a716-446655440000"
              method: "GET"
              rel: "topic"
            - href: "api/v1/votes/idea/660e8400-e29b-41d4-a716-446655440000"
              method: "GET"
              rel: "votes"
            - href: "api/v1/ideas/660e8400-e29b-41d4-a716-446655440000"
              method: "PUT"
              rel: "update"
            - href: "api/v1/ideas/660e8400-e29b-41d4-a716-446655440000"
              method: "DELETE"
              rel: "delete"
            - href: "api/v1/votes"
              method: "POST"
              rel: "vote"
...
paths:
...
  /api/v1/topics:
    get:
      tags: [Topics]
      summary: Get all topics (MySQL)
      description: Public endpoint. Supports pagination, filtering, and sorting.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: status
          in: query
          schema: { type: string, enum: [OPEN, CLOSED] }
          description: Filter by status
        - name: ownerId
          in: query
          schema: { type: string }
          description: Filter by owner user ID
        - name: sortBy
          in: query
          schema: { type: string, enum: [createdAt, title, updatedAt], default: createdAt }
          description: Field to sort by
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Paged list of topics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedTopicResponse' }
              examples:
                ok:
                  summary: Paged topics response
                  value:
                    data:
                      - id: "550e8400-e29b-41d4-a716-446655440000"
                        title: "Dashboard improvements"
                        description: "Proposals to optimize the interface"
                        status: "OPEN"
                        ownerId: "550e8400-e29b-41d4-a716-446655440001"
                        createdAt: "2026-03-30T10:00:00Z"
                        updatedAt: "2026-03-30T15:30:00Z"
                        winningIdea: null
                        _links:
                          - href: "api/v1/topics/550e8400-e29b-41d4-a716-446655440000"
                            method: "GET"
                            rel: "self"
                          - href: "api/v1/ideas/topic/550e8400-e29b-41d4-a716-446655440000"
                            method: "GET"
                            rel: "ideas"
                          - href: "api/v1/topics/550e8400-e29b-41d4-a716-446655440000"
                            method: "PUT"
                            rel: "update"
                          - href: "api/v1/topics/550e8400-e29b-41d4-a716-446655440000"
                            method: "DELETE"
                            rel: "delete"
                    currentPage: 0
                    pageSize: 10
                    totalItems: 1
                    totalPages: 1
        '400': { $ref: '#/components/responses/BadRequest' }
    post:
      tags: [Topics]
      summary: Create a topic (MySQL)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTopicRequest' }
            examples:
              create:
                summary: Create topic request
                value:
                  title: "New mobile features"
                  description: "Ideas for mobile usability improvements"
      responses:
        '201':
          description: Topic created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TopicResponse' }
              examples:
                created:
                  summary: Created topic
                  value:
                    id: "550e8400-e29b-41d4-a716-446655440000"
                    title: "New mobile features"
                    description: "Ideas for mobile usability improvements"
                    status: "OPEN"
                    ownerId: "550e8400-e29b-41d4-a716-446655440001"
                    createdAt: "2026-03-30T10:00:00Z"
                    updatedAt: "2026-03-30T10:00:00Z"
                    winningIdea: null
                    _links:
                      - href: "api/v1/topics/550e8400-e29b-41d4-a716-446655440000"
                        method: "GET"
                        rel: "self"
                      - href: "api/v1/ideas/topic/550e8400-e29b-41d4-a716-446655440000"
                        method: "GET"
                        rel: "ideas"
                      - href: "api/v1/topics/550e8400-e29b-41d4-a716-446655440000"
                        method: "PUT"
                        rel: "update"
                      - href: "api/v1/topics/550e8400-e29b-41d4-a716-446655440000"
                        method: "DELETE"
                        rel: "delete"
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/topics/{id}:
    get:
      tags: [Topics]
      summary: Get topic by ID (MySQL)
      parameters:
        - $ref: '#/components/parameters/TopicId'
      responses:
        '200':
          description: Topic found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TopicResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Topics]
      summary: Update a topic (owner only, MySQL)
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TopicId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTopicRequest' }
            examples:
              close_topic:
                summary: Close a topic (cannot be reopened)
                value:
                  title: "New mobile features - Updated"
                  description: "Focus on UX/UI"
                  status: "CLOSED"
      responses:
        '200':
          description: Topic updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TopicResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Topics]
      summary: Delete a topic (owner only, MySQL)
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TopicId'
      responses:
        '200':
          description: Topic deleted (cascade deletes ideas and votes)
          content:
            application/json:
              examples:
                deleted:
                  summary: Deletion confirmation
                  value:
                    message: "Topic deleted. This action also deleted all related ideas and votes."
                    topicId: "550e8400-e29b-41d4-a716-446655440000"
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ----- IDEAS (V1) -----
...
  /api/v1/ideas/topic/{topicId}:
    get:
      tags: [Ideas]
      summary: Get ideas by topic (MySQL)
      parameters:
        - name: topicId
          in: path
          required: true
          schema: { type: string }
          description: Topic ID (UUID string)
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: List of ideas for the topic (empty array if none)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IdeaResponse' }

  # ----- VOTES (V1) -----
...
  /api/v2/topics:
    get:
      tags: [Topics]
      summary: Get all topics (MongoDB)
      description: Public endpoint. Supports pagination, filtering, and sorting.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: status
          in: query
          schema: { type: string, enum: [OPEN, CLOSED] }
          description: Filter by status
        - name: ownerId
          in: query
          schema: { type: string }
          description: Filter by owner user ID
        - name: sortBy
          in: query
          schema: { type: string, enum: [createdAt, title, updatedAt], default: createdAt }
          description: Field to sort by
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Paged list of topics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedTopicResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
    post:
      tags: [Topics]
      summary: Create a topic (MongoDB)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTopicRequest' }
      responses:
        '201':
          description: Topic created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TopicResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v2/topics/{id}:
    get:
      tags: [Topics]
      summary: Get topic by ID (MongoDB)
      parameters:
        - $ref: '#/components/parameters/TopicId'
      responses:
        '200':
          description: Topic found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TopicResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Topics]
      summary: Update a topic (owner only, MongoDB)
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TopicId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTopicRequest' }
      responses:
        '200':
          description: Topic updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TopicResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Topics]
      summary: Delete a topic (owner only, MongoDB)
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TopicId'
      responses:
        '200':
          description: Topic deleted (cascade deletes ideas and votes)
          content:
            application/json:
              examples:
                deleted:
                  value:
                    message: "Topic deleted. This action also deleted all related ideas and votes."
                    topicId: "550e8400-e29b-41d4-a716-446655440000"
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ----- IDEAS (V2) -----
...
  /api/v2/ideas/topic/{topicId}:
    get:
      tags: [Ideas]
      summary: Get ideas by topic (MongoDB)
      parameters:
        - name: topicId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: List of ideas for the topic (empty array if none)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IdeaResponse' }

  # ----- VOTES (V2) -----
...